# backend/src/main/java/com/linkz/reservation/auth/UserProfileResponse.java
package com.linkz.reservation.auth;

import java.time.LocalDateTime;

public record UserProfileResponse(
        Long id,
        String email,
        LocalDateTime createdAt,
        LocalDateTime updatedAt
) {}