# frontend/api-docs.json
{"openapi":"3.0.1","info":{"title":"Seat Reservation API","description":"API for managing seat reservations","contact":{"name":"Linkz","url":"https://linkz.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearer-jwt":[]}],"paths":{"/api/webhooks/payment-success":{"post":{"tags":["webhook-controller"],"operationId":"handlePaymentSuccess","parameters":[{"name":"X-Signature","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/webhooks/payment-failure":{"post":{"tags":["webhook-controller"],"operationId":"handlePaymentFailure","parameters":[{"name":"X-Signature","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/reservations":{"get":{"tags":["reservation-controller"],"operationId":"getUserReservations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"post":{"tags":["reservation-controller"],"operationId":"reserveSeat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/reservations/{reservationId}/expire":{"post":{"tags":["reservation-controller"],"operationId":"manuallyExpireReservation","parameters":[{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/reservations/{reservationId}/confirm":{"post":{"tags":["reservation-controller"],"operationId":"confirmReservation","parameters":[{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/payments":{"post":{"tags":["payment-controller"],"operationId":"initiatePayment","parameters":[{"name":"reservationId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register_1","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login_1","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/seats":{"get":{"tags":["seat-controller"],"operationId":"getSeats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponse"}}}}}}}},"/api/reservations/{reservationId}":{"get":{"tags":["reservation-controller"],"operationId":"getReservationDetails","parameters":[{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/payments/{paymentId}":{"get":{"tags":["payment-controller"],"operationId":"getPayment","parameters":[{"name":"paymentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["payment-controller"],"operationId":"completePayment","parameters":[{"name":"paymentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentCompletionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}}}},"components":{"schemas":{"ReservationRequest":{"required":["seatId"],"type":"object","properties":{"seatId":{"type":"integer","format":"int64"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"maxLength":255,"minLength":0,"type":"string"},"password":{"maxLength":128,"minLength":6,"type":"string"}}},"SeatResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"seatNumber":{"type":"string"},"status":{"type":"string"}}},"ReservationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"seat_id":{"type":"integer","format":"int64"},"seat_number":{"type":"string"},"created_at":{"type":"string"},"confirmed_at":{"type":"string"},"expired_at":{"type":"string"}}},"PaymentCompletionRequest":{"required":["result"],"type":"object","properties":{"result":{"type":"string"}}},"PaymentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"reservation_id":{"type":"integer","format":"int64"},"status":{"type":"string"},"provider_reference":{"type":"string"},"amount":{"type":"string"},"created_at":{"type":"string"}}},"UserProfileResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ErrorResponse":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"}}}},"securitySchemes":{"bearer-jwt":{"type":"http","description":"JWT Bearer Token","scheme":"bearer","bearerFormat":"JWT"}}}}